requirements.txt
# Flask核心
Flask==2.3.3
Werkzeug==2.3.7

# 数据库
Flask-SQLAlchemy==3.0.5
SQLAlchemy==2.0.20

# 微信开发
wechatpy==1.8.18
pycryptodome==3.18.0

# HTTP请求
requests==2.31.0

# YAML解析
PyYAML==6.0.1

# 定时任务
APScheduler==3.10.4

# 日期时间处理
python-dateutil==2.8.2

# OpenAI SDK (用于调用大模型)
openai==1.3.0

# 环境变量管理
python-dotenv==1.0.0

# 日志
colorlog==6.7.0

